Do not remove nodes if there's no replacement

+++ src/kad.rs
@@ -148,9 +148,13 @@ impl Table {
 
     /// Remove node from the bucket
     #[instrument(skip(self, node), fields(node = &*node.to_string()))]
-    pub fn remove(&mut self, node: NodeId) -> bool {
+    pub fn remove(&mut self, node: NodeId) {
         trace!("Removing from bucket");
         if let Some(bucket) = self.bucket_mut(node) {
+            if bucket.replacements.is_empty() {
+                return;
+            }
+
             for i in 0..bucket.bucket.len() {
                 if bucket.bucket[i].id == node {
                     bucket.bucket.remove(i);
@@ -158,12 +162,10 @@ impl Table {
                         bucket.bucket.push_back(node);
                     }
 
-                    return true;
+                    return;
                 }
             }
         }
-
-        false
     }
 
     pub fn neighbours(&self, peer: NodeId) -> Option<NodeBucket> {
